Unit test failure; refactoring
