+++ ts/packages/knowledgeProcessor/src/conversation/index.ts
@@ -22,6 +22,7 @@ export * from "./actions.js";
 export * from "./searchResponse.js";
 export * from "./searchProcessor.js";
 export * from "./answerSchema.js";
+export * from "./answerContext.js";
 export * from "./answerGenerator.js";
 
 export * as testData from "./testData.js";
